+++ src/main/java/ee/carlrobert/codegpt/actions/editor/AskQuestionAction.java
@@ -2,13 +2,13 @@ package ee.carlrobert.codegpt.actions.editor;
 
 import static java.lang.String.format;
 
-import com.intellij.icons.AllIcons;
 import com.intellij.openapi.editor.Editor;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.ui.DialogWrapper;
 import com.intellij.util.ui.FormBuilder;
 import com.intellij.util.ui.JBUI;
 import com.intellij.util.ui.UI;
+import ee.carlrobert.codegpt.Icons;
 import ee.carlrobert.codegpt.conversations.message.Message;
 import ee.carlrobert.codegpt.toolwindow.chat.ChatToolWindowContentManager;
 import ee.carlrobert.codegpt.ui.UIUtil;
@@ -20,13 +20,12 @@ import javax.swing.JTextArea;
 import javax.swing.SwingUtilities;
 import org.jetbrains.annotations.Nullable;
 
-public class CustomPromptAction extends BaseEditorAction {
+public class AskQuestionAction extends BaseEditorAction {
 
   private static String previousUserPrompt = "";
 
-  CustomPromptAction() {
-    super("Custom Prompt", "Custom prompt description", AllIcons.Actions.Run_anything);
-    EditorActionsUtil.registerAction(this);
+  AskQuestionAction() {
+    super(Icons.QuestionMark);
   }
 
   @Override