+++ src/main/java/ee/carlrobert/codegpt/toolwindow/chat/ChatToolWindowTabPanel.java
@@ -7,6 +7,8 @@ import static java.lang.String.format;
 import com.intellij.openapi.Disposable;
 import com.intellij.openapi.application.ApplicationManager;
 import com.intellij.openapi.diagnostic.Logger;
+import com.intellij.openapi.editor.Editor;
+import com.intellij.openapi.editor.SelectionModel;
 import com.intellij.openapi.project.Project;
 import com.intellij.ui.JBColor;
 import com.intellij.util.ui.JBUI;
@@ -21,6 +23,7 @@ import ee.carlrobert.codegpt.conversations.Conversation;
 import ee.carlrobert.codegpt.conversations.ConversationService;
 import ee.carlrobert.codegpt.conversations.message.Message;
 import ee.carlrobert.codegpt.telemetry.TelemetryAction;
+import ee.carlrobert.codegpt.toolwindow.chat.actionprocessor.ActionProcessorFactory;
 import ee.carlrobert.codegpt.toolwindow.chat.ui.ChatMessageResponseBody;
 import ee.carlrobert.codegpt.toolwindow.chat.ui.ChatToolWindowScrollablePanel;
 import ee.carlrobert.codegpt.toolwindow.chat.ui.ResponsePanel;
@@ -31,10 +34,6 @@ import ee.carlrobert.codegpt.toolwindow.ui.ChatToolWindowLandingPanel;
 import ee.carlrobert.codegpt.ui.OverlayUtil;
 import ee.carlrobert.codegpt.ui.textarea.AppliedActionInlay;
 import ee.carlrobert.codegpt.ui.textarea.UserInputPanel;
-import ee.carlrobert.codegpt.ui.textarea.suggestion.item.CreateDocumentationActionItem;
-import ee.carlrobert.codegpt.ui.textarea.suggestion.item.DocumentationActionItem;
-import ee.carlrobert.codegpt.ui.textarea.suggestion.item.PersonaActionItem;
-import ee.carlrobert.codegpt.ui.textarea.suggestion.item.WebSearchActionItem;
 import ee.carlrobert.codegpt.util.EditorUtil;
 import ee.carlrobert.codegpt.util.file.FileUtil;
 import java.awt.BorderLayout;
@@ -266,42 +265,56 @@ public class ChatToolWindowTabPanel implements Disposable {
     requestHandler.call(callParameters);
   }
 
-  private Unit handleSubmit(String text, List<AppliedActionInlay> appliedInlayActions) {
+  private String processEditorSelection(Editor editor, Message message) {
+    if (editor == null) {
+      return null;
+    }
+
+    SelectionModel selectionModel = editor.getSelectionModel();
+    String selectedText = selectionModel.getSelectedText();
+    if (selectedText == null || selectedText.isEmpty()) {
+      return null;
+    }
+
+    String fileExtension = FileUtil.getFileExtension(editor.getVirtualFile().getName());
+    message.setPrompt(
+        message.getPrompt() + String.format("%n```%s%n%s%n```", fileExtension, selectedText));
+    selectionModel.removeSelection();
+    return selectedText;
+  }
+
+  private Unit handleSubmit(String text, List<? extends AppliedActionInlay> appliedInlayActions) {
     var message = new Message(text);
     var editor = EditorUtil.getSelectedEditor(project);
-    String highlightedText = null;
+
+    var remainingText = new StringBuilder(text);
+    var promptBuilder = new StringBuilder();
+
+    for (var actionInlay : appliedInlayActions) {
+      var inlayOffset = actionInlay.getInlay().getOffset();
+      promptBuilder.append(remainingText, 0, Math.min(inlayOffset, remainingText.length()))
+          .append("\n");
+      ActionProcessorFactory.getProcessor(actionInlay)
+          .process(message, actionInlay, editor, promptBuilder);
+      remainingText.delete(0, inlayOffset);
+    }
+    promptBuilder.append(remainingText);
+
+    String highlightedTextMd = "";
     if (editor != null) {
       var selectionModel = editor.getSelectionModel();
       var selectedText = selectionModel.getSelectedText();
       if (selectedText != null && !selectedText.isEmpty()) {
         var fileExtension = FileUtil.getFileExtension(editor.getVirtualFile().getName());
-        message = new Message(text + format("%n```%s%n%s%n```", fileExtension, selectedText));
-        highlightedText = selectedText;
+        highlightedTextMd = format("\n```%s\n%s\n```\n", fileExtension, selectedText);
         selectionModel.removeSelection();
       }
     }
-    message.setUserMessage(text);
-    message.setWebSearchIncluded(appliedInlayActions.stream()
-        .anyMatch(it -> it.getSuggestion() instanceof WebSearchActionItem));
 
-    var addedDocumentation = CodeGPTKeys.ADDED_DOCUMENTATION.get(project);
-    var appliedInlayExists = appliedInlayActions.stream()
-        .anyMatch(it -> it.getSuggestion() instanceof DocumentationActionItem
-            || it.getSuggestion() instanceof CreateDocumentationActionItem);
-    if (addedDocumentation != null && appliedInlayExists) {
-      message.setDocumentationDetails(addedDocumentation);
-      CodeGPTKeys.ADDED_DOCUMENTATION.set(project, null);
-    }
+    message.setUserMessage(highlightedTextMd + promptBuilder);
+    message.setPrompt(highlightedTextMd + promptBuilder);
 
-    var addedPersona = CodeGPTKeys.ADDED_PERSONA.get(project);
-    var personaInlayExists = appliedInlayActions.stream()
-        .anyMatch(it -> it.getSuggestion() instanceof PersonaActionItem);
-    if (addedPersona != null && personaInlayExists) {
-      message.setPersonaDetails(addedPersona);
-      CodeGPTKeys.ADDED_PERSONA.set(project, null);
-    }
-
-    sendMessage(message, ConversationType.DEFAULT, highlightedText);
+    sendMessage(message, ConversationType.DEFAULT, processEditorSelection(editor, message));
     return Unit.INSTANCE;
   }
 
@@ -374,5 +387,9 @@ public class ChatToolWindowTabPanel implements Disposable {
         BorderLayout.CENTER);
     rootPanel.add(createUserPromptPanel(), BorderLayout.SOUTH);
     return rootPanel;
+  }
+
+  public void addSelection(String fileName, SelectionModel selectionModel) {
+    userInputPanel.addSelection(fileName, selectionModel);
   }
 }