+++ src/main/kotlin/ee/carlrobert/codegpt/settings/service/codegpt/CodeGPTServiceForm.kt
@@ -26,8 +26,9 @@ class CodeGPTServiceForm {
 
     private val chatCompletionModelComboBox =
         ComboBox(ListComboBoxModel(CodeGPTAvailableModels.ALL_CHAT_MODELS)).apply {
-            selectedItem =
-                CodeGPTAvailableModels.findByCode(service<CodeGPTServiceSettings>().state.chatCompletionSettings.model)
+            val chatModel = service<CodeGPTServiceSettings>().state.chatCompletionSettings.model
+            selectedItem = CodeGPTAvailableModels.findByCode(chatModel)
+                ?: CodeGPTAvailableModels.DEFAULT_CHAT_MODEL
             renderer = CustomComboBoxRenderer()
         }
 
@@ -37,9 +38,10 @@ class CodeGPTServiceForm {
     )
 
     private val codeCompletionModelComboBox =
-        ComboBox(ListComboBoxModel(CodeGPTAvailableModels.CODE_MODELS)).apply {
-            selectedItem =
-                CodeGPTAvailableModels.findByCode(service<CodeGPTServiceSettings>().state.codeCompletionSettings.model)
+        ComboBox(ListComboBoxModel(CodeGPTAvailableModels.ALL_CODE_MODELS)).apply {
+            val codeModel = service<CodeGPTServiceSettings>().state.codeCompletionSettings.model
+            selectedItem = CodeGPTAvailableModels.findByCode(codeModel)
+                ?: CodeGPTAvailableModels.DEFAULT_CODE_MODEL
             renderer = CustomComboBoxRenderer()
         }