+++ CHANGELOG.md
@@ -6,6 +6,30 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- New `o1-preview` and `o1-mini` models for CodeGPT users 
+
+## [2.11.1-241.1] - 2024-09-12
+
+### Added
+
+- Git history referencing in prompts
+- Quick code snippet referencing in prompts [#690](https://github.com/carlrobertoh/CodeGPT/issues/690)
+- Name suggestions for languages other than Java [#615](https://github.com/carlrobertoh/CodeGPT/issues/615)
+
+### Fixed
+
+- Prompt text field persona suggestions [#685](https://github.com/carlrobertoh/CodeGPT/issues/685)
+- Text field autofocus when opening the chat toolwindow [#691](https://github.com/carlrobertoh/CodeGPT/issues/691)
+- Git commit dialog selected changes not being included in diff
+- Ollama server host configuration overriding [#684](https://github.com/carlrobertoh/CodeGPT/issues/684)
+
+### Improved
+
+- Model combobox menu UI/UX by extracting the providers into standalone subgroups [#681](https://github.com/carlrobertoh/CodeGPT/pull/681)
+- Prompt text field inlay rendering
+
 ## [2.11.0-241.1] - 2024-09-05
 
 ### Added
@@ -627,7 +651,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `OPENAI_API_KEY` persistence, key is saved in the OS password safe from now on
 
-[Unreleased]: https://github.com/carlrobertoh/CodeGPT/compare/v2.11.0-241.1...HEAD
+[Unreleased]: https://github.com/carlrobertoh/CodeGPT/compare/v2.11.1-241.1...HEAD
+[2.11.1-241.1]: https://github.com/carlrobertoh/CodeGPT/compare/v2.11.0-241.1...v2.11.1-241.1
 [2.11.0-241.1]: https://github.com/carlrobertoh/CodeGPT/compare/v2.10.2-241.1...v2.11.0-241.1
 [2.10.2-241.1]: https://github.com/carlrobertoh/CodeGPT/compare/v2.10.1-241.1...v2.10.2-241.1
 [2.10.1-241.1]: https://github.com/carlrobertoh/CodeGPT/compare/v2.10.0-233...v2.10.1-241.1